ChaCha20 cleanup and performance improvements (#117)
* chacha20: Optimizations for processing keystream blocks

* chacha20: Remove Result return on xor_keystream_block

All errors that could be returned by xor_keystream_block were only possible
if the ChaCha20 implementation didn't handle blocks correctly, so a user
would never able to trigger these themselves.

We replace the checks with debug_assert! and tests conditionally when the build
is debug.

* chacha20: Remove Option<> on block counter and instead just ignore it if state is used for HChaCha

* chacha20: Make round functions macros yielding small amount of better throughput

* chacha20: Optimize processing keystream going from ~515 MiB/s throughput to ~603 MiB/s

- XOR arrays of blocksize length in encrypt_in_place instead of XORing each individual row.
- Use Zeroizing::new()n on tmp keystream_block array, wiping it first when we are done processing
the all keystream. Also, it ensured it gets dropped if we return Err() on overflowing counter.

* chacha20: Bring back RFC8439 key schedule test