Add ability to randomly generate SecretKey and Nonce in (StreamCipher/Aead)TestRunner
This must be implemented seperately from the test runner, and is therefor a pub function.
The reason for this is, that the trait that let's us randomly generate these types, should
only be available when testing. But since StreamCipherTestRunner is used in intergation tests (/tests),
compiling it conditionally with cfg(test) would still make it inaccessible in /tests. Using
cfg(debug_assertions) is not desired either, because we want to test in both release and debug mode.