Use encoded parameters in hash_password_verify
The `pwhash::hash_password_verify` function is central to the UX of the
password verification portion of this library. Currently, it requires callers
to specify the iterations and memory to be used in the Argon2i hash that the
function performs.

However, one of the arguments to the function is a `PasswordHash` type. This
type represents a password that has already been hashed, and it contains the
memory and iterations that were used to hash it. It's almost certainly
a logical error to hash a password for the purpose of comparing it to another
hash generated using different parameters. We can save users from having to
specify the paramters---and eventually making a mistake doing so---by having
`hash_password_verify` just pull the parameters from the given `PasswordHash`.
We retain the original functionality in a separate function called
`hash_password_verify_with`.