clarify and fix hash_password_verify docs example
Before this commit, the example in the docs for the hash_password_verify
function included an erroneous assertion that calculating two hashes with
different parameters from the same password and then verifying those hashes
against that same password using hash_password_verify would result in an error.
In fact, the entire reason for this PR is to make that process result in
a successful verification.

I also added an assertion that shows how using the wrong password could result
in a failure to verify the password.

Finally, I added a comment that clarifies why both hashes verify correctly, and
that this is the intended behavior.