Use encoded parameters in hash_password_verify (#137)
* Use encoded parameters in hash_password_verify

The `pwhash::hash_password_verify` function is central to the UX of the
password verification portion of this library. Currently, it requires callers
to specify the iterations and memory to be used in the Argon2i hash that the
function performs.

However, one of the arguments to the function is a `PasswordHash` type. This
type represents a password that has already been hashed, and it contains the
memory and iterations that were used to hash it. It's almost certainly
a logical error to hash a password for the purpose of comparing it to another
hash generated using different parameters. We can save users from having to
specify the paramters---and eventually making a mistake doing so---by having
`hash_password_verify` just pull the parameters from the given `PasswordHash`.
We retain the original functionality in a separate function called
`hash_password_verify_with`.

* Use code blocks for variables in docs

* Fix typo in docs

* Remove hash_password_verify_with and its tests

It's likely rare that users will need this functionality. If anyone
needs it, they can use types in the `hazardous` module to achieve exactly what
`hash_password_verify_with` provides. Removing the high-level interface also
has the benefit of not leading users astray who don't know what they're looking
for, and may (likely incorrectly) choose `hash_password_verify_with` over
`hash_password_verify`.

* Remove reference to nonexistant function

* clarify and fix hash_password_verify docs example

Before this commit, the example in the docs for the hash_password_verify
function included an erroneous assertion that calculating two hashes with
different parameters from the same password and then verifying those hashes
against that same password using hash_password_verify would result in an error.
In fact, the entire reason for this PR is to make that process result in
a successful verification.

I also added an assertion that shows how using the wrong password could result
in a failure to verify the password.

Finally, I added a comment that clarifies why both hashes verify correctly, and
that this is the intended behavior.

* test that modifying encoded pwhash causes failure

This commit adds three tests that each show how changing the parameters
within the encoded password hash string will cause a verification
failure, as is the intended behavior. The three tests check for
modifying the memory, iterations, and then memory and iterations
together.

Co-authored-by: Vincent Mutolo <vince@mutolo.org>