switch to crates.io version of fiat-crypto
Previously, we vendored the generated code from the fiat-crypto
project. It was brought to our attention that the officially
published fiat-crypto on crates.io is sufficiently up-to-date
to use in Orion and that there's no need to manually vendor
the code.

Additionally, using the crates.io version integrates well
with Dependabot so we can get alerts when our version of fiat-crypto
is out of date. This was a challenge we would have had to solve
with the vendored code that we now get solved for free.