convert blake2b::Digest to const generics
This is still a WIP. The basic method was to define an inner type
that behaves the way we want it to, e.g. with special Drop impls,
etc. Then for each newtype we want to create, we create a small
wrapper over the inner type, like `struct Digest(Blake2bArray)`,
and implement two small traits, essentially From and AsRef, to
convert to that inner type.

In another module, we provide more complex blanket implementations
for any types that implement From and AsRef for our inner type.
This allow us to, for example, auto-implement `len`, `from_slice`,
etc for the newtypes without exposing the inner type (which could
be hazardous for inner types like the planned `PrivateArray` type.