split Data into PublicData and SecretData
Because marker traits aren't mutually exclusive (as far as the
type system is concerned), they didn't end up being a good way
to represent whether Data was public or private info.

This commit moves to using two separate data container types:
one for public info, and one for private. The upside is that
this is significantly simpler than using marker traits, and that
the PublicData type can be documented separately from the SecretData
type. The downside is that we duplicated a bunch of code since
PublicData and SecretData are basically identical with the exception
of Debug, PartialEq, and the AsRef<[u8]>/unprotected_as_bytes split.